server/requirements.txt
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.2.0
chromadb==0.4.24
python-multipart==0.0.9
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
python-dotenv==1.0.1
tiktoken==0.7.0
sentence-transformers==3.0.0
numpy==1.26.4
pandas==2.2.2
